Eliminate warnings on Visual Studio 10 W64. On Win64, the size of 'int' is stil 4 bytes and the size of 'size_t' is 8 bytes. A conversion from 'size_t' to 'int', causes warning with compiling.
SVN-Revision: 1772

+++ libarchive/archive_read_support_compression_program.c
@@ -205,7 +205,7 @@ program_bidder_bid(struct archive_read_filter_bidder *self,
 		/* No match, so don't bid. */
 		if (memcmp(p, state->signature, state->signature_len) != 0)
 			return (0);
-		return (state->signature_len * 8);
+		return ((int)state->signature_len * 8);
 	}
 
 	/* Otherwise, bid once and then never bid again. */
